preparing for terraform lambda
