+++ deployment/webservice-lambda/Dockerfile
@@ -7,9 +7,7 @@ COPY [ "Pipfile", "Pipfile.lock", "./" ]
 
 RUN pipenv install --system --deploy
 
-COPY [ "lgb_preds.parquet" , "lgb_preds.parquet" ]
-COPY [ "items.parquet" , "items.parquet" ]
-
-COPY [ "lambda_function.py", "lambda_function.py" ]
+# COPY [ "lgb_preds.parquet" , "lgb_preds.parquet" ]
+COPY [ "prep.py", "items.parquet" , "lambda_function.py", "./" ]
 
 CMD [ "lambda_function.lambda_handler" ]
