+++ deployment/webservice-lambda/Pipfile
@@ -5,12 +5,12 @@ name = "pypi"
 
 [packages]
 pandas = "*"
-flask = "*"
 pyarrow = "*"
-gunicorn = "*"
 
 [dev-packages]
 requests = "*"
+mlflow = "*"
+boto3 = "*"
 
 [requires]
 python_version = "3.9"
